+++ pkg/deployment/resources/inspector/at_gvk.go
@@ -23,22 +23,21 @@ package inspector
 import (
 	"k8s.io/apimachinery/pkg/runtime/schema"
 
-	"github.com/arangodb/kube-arangodb/pkg/util/k8sutil/inspector/anonymous"
 	"github.com/arangodb/kube-arangodb/pkg/util/k8sutil/inspector/constants"
 )
 
-func (p *podsInspector) Anonymous(gvk schema.GroupVersionKind) (anonymous.Interface, bool) {
-	g := constants.PodGK()
+func (p *arangoTasksInspectorV1) GroupVersionKind() schema.GroupVersionKind {
+	return constants.ArangoTaskGKv1()
+}
 
-	if g.Kind == gvk.Kind && g.Group == gvk.Group {
-		switch gvk.Version {
-		case constants.PodVersionV1, DefaultVersion:
-			if p.v1 == nil || p.v1.err != nil {
-				return nil, false
-			}
-			return &podsInspectorAnonymousV1{i: p.state}, true
-		}
-	}
+func (p *arangoTasksInspectorV1) GroupVersionResource() schema.GroupVersionResource {
+	return constants.ArangoTaskGRv1()
+}
 
-	return nil, false
+func (p *arangoTasksInspector) GroupKind() schema.GroupKind {
+	return constants.ArangoTaskGK()
+}
+
+func (p *arangoTasksInspector) GroupResource() schema.GroupResource {
+	return constants.ArangoTaskGR()
 }