+++ pkg/deployment/resources/inspector/nodes_anonymous_v1.go
@@ -25,6 +25,8 @@ import (
 
 	meta "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/apimachinery/pkg/types"
+
+	"github.com/arangodb/kube-arangodb/pkg/util/k8sutil/inspector/constants"
 )
 
 type nodesInspectorAnonymousV1 struct {
@@ -36,21 +38,21 @@ func (e *nodesInspectorAnonymousV1) Get(ctx context.Context, name string, opts m
 }
 
 func (e *nodesInspectorAnonymousV1) Create(ctx context.Context, obj meta.Object, opts meta.CreateOptions) (meta.Object, error) {
-	return nil, newNotImplementedError(NodeGKv1())
+	return nil, newNotImplementedError(constants.NodeGKv1())
 }
 
 func (e *nodesInspectorAnonymousV1) Update(ctx context.Context, obj meta.Object, opts meta.UpdateOptions) (meta.Object, error) {
-	return nil, newNotImplementedError(NodeGKv1())
+	return nil, newNotImplementedError(constants.NodeGKv1())
 }
 
 func (e *nodesInspectorAnonymousV1) UpdateStatus(ctx context.Context, obj meta.Object, opts meta.UpdateOptions) (meta.Object, error) {
-	return nil, newNotImplementedError(NodeGKv1())
+	return nil, newNotImplementedError(constants.NodeGKv1())
 }
 
 func (e *nodesInspectorAnonymousV1) Patch(ctx context.Context, name string, pt types.PatchType, data []byte, opts meta.PatchOptions, subresources ...string) (result meta.Object, err error) {
-	return nil, newNotImplementedError(NodeGKv1())
+	return nil, newNotImplementedError(constants.NodeGKv1())
 }
 
 func (e *nodesInspectorAnonymousV1) Delete(ctx context.Context, name string, opts meta.DeleteOptions) error {
-	return newNotImplementedError(NodeGKv1())
+	return newNotImplementedError(constants.NodeGKv1())
 }